requirements.txt
Django==1.9.6
django-appconf==1.0.2
django-autoslug==1.9.3
django-ckeditor==5.0.3
django-imagekit==3.3
django-sass-processor==0.3.5
libsass==0.11.1
pilkit==1.1.13
Pillow==3.2.0
six==1.10.0
